remove redundant mod
